completely redo indexing to not rely on c++17 recursive iterator.
it would invalidate the iterator if a directory was encountered with flags or other security restrictions.  there didn't seem to be a way to continue processing with that iterator as it gets set to end on the exception.  the directory_iterator() also throws an error, but can be continued just fine.  alas, had to add cyclic path detection which undoubtedly added some overhead, but works well.